update local master

+++ scripts/build/build/__init__.py
@@ -4,9 +4,8 @@ import shutil
 from enum import Enum, auto
 from typing import Sequence
 
-from .targets import ALL, Target
-
-ALL_TARGETS = ALL
+from .targets import BUILD_TARGETS
+from builders.builder import BuilderOptions
 
 
 class BuildSteps(Enum):
@@ -26,8 +25,7 @@ class Context:
         self.output_prefix = output_prefix
         self.completed_steps = set()
 
-    def SetupBuilders(self, targets: Sequence[Target],
-                      enable_flashbundle: bool):
+    def SetupBuilders(self, targets: Sequence[str], options: BuilderOptions):
         """
         Configures internal builders for the given platform/board/app
         combination.
@@ -35,9 +33,16 @@ class Context:
 
         self.builders = []
         for target in targets:
-            self.builders.append(target.Create(
-                self.runner, self.repository_path, self.output_prefix,
-                enable_flashbundle))
+            found = False
+            for choice in BUILD_TARGETS:
+                builder = choice.Create(target, self.runner, self.repository_path,
+                                        self.output_prefix, options)
+                if builder:
+                    self.builders.append(builder)
+                    found = True
+
+            if not found:
+                logging.error(f"Target '{target}' could not be found. Nothing executed for it")
 
         # whenever builders change, assume generation is required again
         self.completed_steps.discard(BuildSteps.GENERATED)
@@ -68,7 +73,8 @@ class Context:
     def CleanOutputDirectories(self):
         for builder in self.builders:
             logging.warn('Cleaning %s', builder.output_dir)
-            shutil.rmtree(builder.output_dir)
+            if os.path.exists(builder.output_dir):
+                shutil.rmtree(builder.output_dir)
 
         # any generated output was cleaned
         self.completed_steps.discard(BuildSteps.GENERATED)
